refactor: util functions
